refactor: modularize and organize table component
